Reorganize in crates
