+++ providers/provider-memory/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "provider-memory"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+guest-metrics = { path = "../../guest-metrics" }
+futures = "0.3"
+tokio = { version = "1.25.0", features = ["time"] }
