+++ providers/provider-netlink/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "provider-netlink"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+futures = "0.3"
+guest-metrics = { path = "../../guest-metrics" }
+tokio = { version = "1", features = ["rt"] }
+uuid = "1.11"
+anyhow = "*"
+
+netlink-packet-core = { version = "0.7.0" }
+netlink-packet-route = { version = ">=0.18.0, <0.20" }
+netlink-proto = { version = "0.11.2" }
+rtnetlink = { version = "0.14.0" }
