+++ providers/provider-simple/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "provider-simple"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+futures = "0.3"
+guest-metrics = { path = "../../guest-metrics" }
+uuid = { version = "1.11", features = ["v4"] }
+tokio = { version = "1", features = ["time"] }
+
+vif-detect = { path = "../vif-detect" }
+network-interface = "2.0"
