+++ providers/vif-detect/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "vif-detect"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+log = "*"
+guest-metrics = { path = "../../guest-metrics" }
