+++ publishers/publisher-console/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "publisher-console"
+edition = "2021"
+version.workspace = true
+repository.workspace = true
+categories.workspace = true
+
+[dependencies]
+guest-metrics = { path = "../../guest-metrics" }
+uuid = "1.11"
+futures = "0.3"
