--- src/provider/mod.rs
@@ -1,3 +1,0 @@
-pub mod kernel;
-pub mod memory;
-pub mod net;
